# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hommies Group Api Documentation",
        "contact": {
            "name": "developer",
            "email": "rohit7305.rk@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        },
        {
            "url": "http://localhost/hommiesGroup/api",
            "description": "local server"
        },
        {
            "url": "https://hommiesgroup.com/api",
            "description": "Live server"
        }
    ],
    "paths": {
...
        "/property/list": {
            "get": {
                "tags": [
                    "property"
                ],
                "operationId": "properties",
                "responses": {
                    "200": {
                        "description": "Property List",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Property One"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/property/properties": {
            "post": {
                "tags": [
                    "property"
                ],
                "summary": "Add a new property",
                "description": "Register a new property with additional details",
                "operationId": "addProperty",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "no_of_rooms",
                                    "no_of_beds",
                                    "price",
                                    "price_type",
                                    "address",
                                    "latitude",
                                    "longitude",
                                    "bathrooms",
                                    "property_type",
                                    "adult",
                                    "children",
                                    "infants",
                                    "property_id_proof_1",
                                    "property_id_proof_2"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Beautiful House"
                                    },
                                    "no_of_rooms": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "no_of_beds": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "about": {
                                        "type": "string",
                                        "example": "A cozy place for vacation."
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150.75
                                    },
                                    "price_type": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main St, Springfield"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 40.7128
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -74.006
                                    },
                                    "bathrooms": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "property_type": {
                                        "type": "string",
                                        "example": "1,0,2"
                                    },
                                    "adult": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "children": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "infants": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "amenities_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    },
                                    "property_id_proof_1": {
                                        "description": "First property ID proof document",
                                        "type": "file"
                                    },
                                    "property_id_proof_2": {
                                        "description": "Second property ID proof document",
                                        "type": "file"
                                    },
                                    "images[]": {
                                        "description": "Multiple images to upload",
                                        "type": "array",
                                        "items": {
                                            "type": "file"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Property created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Property created successfully."
                                        },
                                        "property": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "name": "Beautiful House",
                                                "no_of_rooms": 3,
                                                "no_of_beds": 2,
                                                "price": 150.75,
                                                "address": "123 Main St, Springfield"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/property/edit/{id}": {
            "post": {
                "tags": [
                    "property"
                ],
                "operationId": "updateProperty",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Property"
                                    },
                                    "no_of_rooms": {
                                        "type": "integer",
                                        "example": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Property updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Property updated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Property not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Property not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/property/myFavourites": {
            "get": {
                "tags": [
                    "property"
                ],
                "summary": "Get a list of favourite items for the authenticated user",
                "description": "Returns a list of active favourite items for the authenticated user.",
                "operationId": "myFavourites",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "No favourite items found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No providers are in your favourites"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}